+++ api/v1/routes.go
@@ -8,6 +8,21 @@ import (
 	"github.com/go-chi/jwtauth"
 )
 
+var (
+	tokenAuth *jwtauth.JWTAuth
+	tokenString string
+)
+
+func init() {
+	tokenAuth = jwtauth.New("HS256", []byte("secret"), nil)
+}
+
+func main() {
+	addr := ":3000/api/v1"
+	log.Printf("Starting server on %v\n", addr)
+	http.ListenAndServe(addr, Router())
+}
+
 func Router() http.Handler {
 	mux := chi.NewRouter()
