+++ api/v1/tests/session_handler_test.go
@@ -9,7 +9,7 @@ import (
 	"strings"
 	"testing"
 	
-	"github.com/motonary/Fortuna/api"
+	api "github.com/motonary/Fortuna/api/v1"
 	"github.com/motonary/Fortuna/entity"
 )
